Change application name to Explosion
