+++ .gitignore
@@ -20,7 +20,7 @@ erl_crash.dump
 *.ez
 
 # Ignore package tarball (built via "mix hex.build").
-live_view_demo-*.tar
+explosion-*.tar
 
 # If NPM crashes, it generates a log, let's ignore it too.
 npm-debug.log
