+++ config/prod.secret.exs
@@ -11,7 +11,7 @@ database_url =
     For example: ecto://USER:PASS@HOST/DATABASE
     """
 
-config :live_view_demo, LiveViewDemo.Repo,
+config :explosion, Explosion.Repo,
   # ssl: true,
   url: database_url,
   pool_size: String.to_integer(System.get_env("POOL_SIZE") || "10")
@@ -23,7 +23,7 @@ secret_key_base =
     You can generate one by calling: mix phx.gen.secret
     """
 
-config :live_view_demo, LiveViewDemoWeb.Endpoint,
+config :explosion, ExplosionWeb.Endpoint,
   http: [:inet6, port: String.to_integer(System.get_env("PORT") || "4000")],
   secret_key_base: secret_key_base
 
@@ -32,7 +32,7 @@ config :live_view_demo, LiveViewDemoWeb.Endpoint,
 # If you are doing OTP releases, you need to instruct Phoenix
 # to start each relevant endpoint:
 #
-#     config :live_view_demo, LiveViewDemoWeb.Endpoint, server: true
+#     config :explosion, ExplosionWeb.Endpoint, server: true
 #
 # Then you can assemble a release by calling `mix release`.
 # See `mix help release` for more information.
