+++ lib/explosion/live_view_demo/repo.ex
@@ -1,0 +1,5 @@
+defmodule Explosion.Repo do
+  use Ecto.Repo,
+    otp_app: :explosion,
+    adapter: Ecto.Adapters.Postgres
+end
