+++ lib/explosion_web/controllers/page_controller.ex
@@ -1,0 +1,7 @@
+defmodule ExplosionWeb.PageController do
+  use ExplosionWeb, :controller
+
+  def index(conn, _params) do
+    render(conn, "index.html")
+  end
+end
