+++ lib/explosion_web/live/clock_live.ex
@@ -1,0 +1,26 @@
+defmodule ExplosionWeb.ClockLive do
+  use Phoenix.LiveView
+  import Calendar.Strftime
+
+  def render(assigns) do
+    ~L"""
+    <div>
+      <h2 phx-click="boom">It's <%= strftime!(@date, "%r") %></h2>
+    </div>
+    """
+  end
+
+  def mount(_session, socket) do
+    if connected?(socket), do: :timer.send_interval(1000, self(), :tick)
+
+    {:ok, put_date(socket)}
+  end
+
+  def handle_info(:tick, socket) do
+    {:noreply, put_date(socket)}
+  end
+
+  defp put_date(socket) do
+    assign(socket, date: :calendar.local_time())
+  end
+end
