+++ lib/explosion_web/router.ex
@@ -1,0 +1,28 @@
+defmodule ExplosionWeb.Router do
+  use ExplosionWeb, :router
+
+  pipeline :browser do
+    plug :accepts, ["html"]
+    plug :fetch_session
+    plug Phoenix.LiveView.Flash
+    plug :fetch_flash
+    plug :protect_from_forgery
+    plug :put_secure_browser_headers
+  end
+
+  pipeline :api do
+    plug :accepts, ["json"]
+  end
+
+  scope "/", ExplosionWeb do
+    pipe_through :browser
+
+    get "/", PageController, :index
+    live "/clock", ClockLive
+  end
+
+  # Other scopes may use custom stacks.
+  # scope "/api", ExplosionWeb do
+  #   pipe_through :api
+  # end
+end
