+++ lib/explosion_web/views/layout_view.ex
@@ -1,0 +1,3 @@
+defmodule ExplosionWeb.LayoutView do
+  use ExplosionWeb, :view
+end
