+++ lib/explosion_web/views/page_view.ex
@@ -1,0 +1,3 @@
+defmodule ExplosionWeb.PageView do
+  use ExplosionWeb, :view
+end
