--- lib/live_view_demo/repo.ex
@@ -1,5 +1,0 @@
-defmodule LiveViewDemo.Repo do
-  use Ecto.Repo,
-    otp_app: :live_view_demo,
-    adapter: Ecto.Adapters.Postgres
-end
