--- lib/live_view_demo_web/controllers/page_controller.ex
@@ -1,7 +1,0 @@
-defmodule LiveViewDemoWeb.PageController do
-  use LiveViewDemoWeb, :controller
-
-  def index(conn, _params) do
-    render(conn, "index.html")
-  end
-end
