--- lib/live_view_demo_web/views/layout_view.ex
@@ -1,3 +1,0 @@
-defmodule LiveViewDemoWeb.LayoutView do
-  use LiveViewDemoWeb, :view
-end
