--- lib/live_view_demo_web/views/page_view.ex
@@ -1,3 +1,0 @@
-defmodule LiveViewDemoWeb.PageView do
-  use LiveViewDemoWeb, :view
-end
