+++ rename.sh
@@ -1,0 +1,12 @@
+#!/bin/bash
+
+CURRENT_OTP="explosion"
+
+NEW_OTP="explosion"
+
+mv lib/$CURRENT_OTP lib/$NEW_OTP
+mv lib/$CURRENT_OTP.ex lib/$NEW_OTP.ex
+mv lib/${CURRENT_OTP}_web lib/${NEW_OTP}_web
+mv lib/${CURRENT_OTP}_web.ex lib/${NEW_OTP}_web.ex
+mv test/$CURRENT_OTP test/$NEW_OTP
+mv test/${CURRENT_OTP}_web test/${NEW_OTP}_web
