+++ test/explosion_web/controllers/page_controller_test.exs
@@ -1,0 +1,8 @@
+defmodule ExplosionWeb.PageControllerTest do
+  use ExplosionWeb.ConnCase
+
+  test "GET /", %{conn: conn} do
+    conn = get(conn, "/")
+    assert html_response(conn, 200) =~ "My Phrenzy Entry"
+  end
+end
