+++ test/explosion_web/views/layout_view_test.exs
@@ -1,0 +1,3 @@
+defmodule ExplosionWeb.LayoutViewTest do
+  use ExplosionWeb.ConnCase, async: true
+end
