+++ test/explosion_web/views/page_view_test.exs
@@ -1,0 +1,3 @@
+defmodule ExplosionWeb.PageViewTest do
+  use ExplosionWeb.ConnCase, async: true
+end
