--- test/live_view_demo_web/controllers/page_controller_test.exs
@@ -1,8 +1,0 @@
-defmodule LiveViewDemoWeb.PageControllerTest do
-  use LiveViewDemoWeb.ConnCase
-
-  test "GET /", %{conn: conn} do
-    conn = get(conn, "/")
-    assert html_response(conn, 200) =~ "My Phrenzy Entry"
-  end
-end
