--- test/live_view_demo_web/views/layout_view_test.exs
@@ -1,3 +1,0 @@
-defmodule LiveViewDemoWeb.LayoutViewTest do
-  use LiveViewDemoWeb.ConnCase, async: true
-end
