--- test/live_view_demo_web/views/page_view_test.exs
@@ -1,3 +1,0 @@
-defmodule LiveViewDemoWeb.PageViewTest do
-  use LiveViewDemoWeb.ConnCase, async: true
-end
